pyproject.toml
[project]
name = "causal-image-embedding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "scikit-learn>=1.8.0",
    "tensorflow>=2.21.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "tqdm>=4.67.3",
    "types-tensorflow>=2.18.0.20260322",
    "types-tqdm>=4.67.3.20260303",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.11",
    "pandas-stubs>=2.0",
    "pytest>=8.0",
    "ruff>=0.8",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra"

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "UP",
    "B",
    "SIM",
    "A",
    "RUF",
]
ignore = [
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"src/dataset.py" = ["SIM108"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
mypy_path = "src"
explicit_package_bases = true
warn_unused_ignores = true
check_untyped_defs = true
disallow_untyped_defs = false
warn_return_any = true
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "hydra.*",
    "matplotlib.*",
    "omegaconf.*",
    "PIL.*",
    "sklearn.*",
    "tensorflow.*",
    "torch.*",
    "torchvision.*",
    "tqdm.*",
]
ignore_missing_imports = true
